refactor: move prism into separate dir
