+++ src/lens/index.ts
@@ -1,2 +1,1 @@
 export * from './lens'
-export * from './prism'
