+++ src/maybe/utils/index.ts
@@ -4,3 +4,4 @@ export * from './from-undefined'
 export * from './merge'
 export * from './or'
 export * from './from-throwable'
+export * from './is'
